Make MetSegaSegmnetFeatureComputation tread safe.
There was the problem that the feature table becomes inconsistent, when
setFeatureSelection is called during the computation of the feature table.
This is solved by removing the setFeatureSelection method. The feature
selection is set within the constructor and must not be changed afterwards.

Instead a new MetaSegSegmentFeatureComputation needs to be created,
whenever the user changes the feature selection.

Add a cancel functionality to cancel the computation of feature rows.